Refactor create journal and update journal page.
