+++ lib/presentation/components/journal_form_date.dart
@@ -1,0 +1,80 @@
+import 'package:flutter/material.dart';
+
+class DateForm extends StatelessWidget {
+  const DateForm(
+      {super.key, required this.datePicked, required this.onDatePicked});
+
+  final DateTime? datePicked;
+  final ValueChanged<DateTime?> onDatePicked;
+
+  TextStyle get textStyle => const TextStyle(
+        fontWeight: FontWeight.bold,
+        fontSize: 18,
+      );
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      width: MediaQuery.sizeOf(context).width / 1.2,
+      decoration: const BoxDecoration(
+        border: Border(
+          bottom: BorderSide(
+            color: Colors.transparent,
+            width: 0.5,
+          ),
+        ),
+      ),
+      child: Stack(
+        children: [
+          Padding(
+            padding: const EdgeInsets.only(left: 10),
+            child: Align(
+              alignment: Alignment.centerLeft,
+              child: DatePickerWrapper(
+                onDatePicked: onDatePicked,
+                child: const Icon(
+                  Icons.calendar_month,
+                  size: 25,
+                ),
+              ),
+            ),
+          ),
+          Align(
+            alignment: Alignment.center,
+            child: DatePickerWrapper(
+              onDatePicked: onDatePicked,
+              child: Text(
+                '${datePicked?.year}. ${datePicked?.month}. ${datePicked?.day}.',
+                style: textStyle,
+              ),
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
+
+class DatePickerWrapper extends StatelessWidget {
+  const DatePickerWrapper(
+      {super.key, required this.child, required this.onDatePicked});
+  final ValueChanged<DateTime?> onDatePicked;
+  final Widget child;
+
+  @override
+  Widget build(BuildContext context) {
+    return GestureDetector(
+      onTap: () async {
+        DateTime? pickedDate = await showDatePicker(
+          context: context,
+          firstDate: DateTime(1980),
+          lastDate: DateTime(2050),
+        );
+        if (pickedDate != null) {
+          onDatePicked(pickedDate);
+        }
+      },
+      child: child,
+    );
+  }
+}
