+++ lib/presentation/pages/page_journal/create_journal.dart
@@ -1,3 +1,6 @@
+import 'package:farmers_journal/presentation/components/journal_form_content.dart';
+import 'package:farmers_journal/presentation/components/journal_form_date.dart';
+import 'package:farmers_journal/presentation/components/journal_form_title.dart';
 import 'package:farmers_journal/presentation/components/layout_images.dart';
 import 'package:farmers_journal/presentation/components/show_snackbar.dart';
 import 'package:farmers_journal/presentation/components/styles/button.dart';
@@ -60,7 +63,7 @@ class _CreateJournalFormState extends ConsumerState<ConsumerStatefulWidget> {
       child: Column(
         spacing: 10,
         children: [
-          _DateForm(
+          DateForm(
             datePicked: date,
             onDatePicked: onDatePicked,
           ),
@@ -75,13 +78,13 @@ class _CreateJournalFormState extends ConsumerState<ConsumerStatefulWidget> {
                   ),
                 )
               : const SizedBox.shrink(),
-          _TitleForm(
+          TitleForm(
             controller: titleController,
             notValid: isFormEmpty,
           ),
           Expanded(
             flex: 3,
-            child: _ContentForm(
+            child: ContentForm(
               controller: contentController,
               onImagePick: () async {
                 List<XFile> _images =
@@ -98,7 +101,6 @@ class _CreateJournalFormState extends ConsumerState<ConsumerStatefulWidget> {
               orElse: () {
                 return () async {
                   _formKey.currentState?.validate();
-
                   if (!isFormEmpty) {
                     _formKey.currentState?.save();
                     try {
@@ -152,215 +154,6 @@ class _CreateJournalFormState extends ConsumerState<ConsumerStatefulWidget> {
                 fontWeight: FontWeight.bold,
                 color: Colors.white,
               ),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-class _DatePickerWrapper extends StatelessWidget {
-  const _DatePickerWrapper(
-      {super.key, required this.child, required this.onDatePicked});
-  final ValueChanged<DateTime?> onDatePicked;
-  final Widget child;
-  @override
-  Widget build(BuildContext context) {
-    return GestureDetector(
-      onTap: () async {
-        DateTime? pickedDate = await showDatePicker(
-          context: context,
-          firstDate: DateTime(1980),
-          lastDate: DateTime(2050),
-        );
-        if (pickedDate != null) {
-          onDatePicked(pickedDate);
-        }
-      },
-      child: child,
-    );
-  }
-}
-
-class _DateForm extends StatelessWidget {
-  const _DateForm(
-      {super.key, required this.datePicked, required this.onDatePicked});
-  final DateTime? datePicked;
-  final ValueChanged<DateTime?> onDatePicked;
-  TextStyle get textStyle => const TextStyle(
-        fontWeight: FontWeight.bold,
-        fontSize: 18,
-      );
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      width: MediaQuery.sizeOf(context).width / 1.2,
-      decoration: const BoxDecoration(
-        border: Border(
-          bottom: BorderSide(
-            color: Colors.transparent,
-            width: 0.5,
-          ),
-        ),
-      ),
-      child: Stack(
-        children: [
-          Padding(
-            padding: const EdgeInsets.only(left: 10),
-            child: Align(
-              alignment: Alignment.centerLeft,
-              child: _DatePickerWrapper(
-                onDatePicked: onDatePicked,
-                child: const Icon(
-                  Icons.calendar_month,
-                  size: 25,
-                ),
-              ),
-            ),
-          ),
-          Align(
-            alignment: Alignment.center,
-            child: _DatePickerWrapper(
-              onDatePicked: onDatePicked,
-              child: Text(
-                '${datePicked?.year}. ${datePicked?.month}. ${datePicked?.day}.',
-                style: textStyle,
-              ),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-class _TitleForm extends StatelessWidget {
-  const _TitleForm(
-      {super.key, required this.controller, required this.notValid});
-
-  final TextEditingController controller;
-  final bool notValid;
-  TextStyle get textStyle => const TextStyle(
-        fontSize: 18,
-        fontWeight: FontWeight.bold,
-      );
-  @override
-  Widget build(BuildContext context) {
-    return SizedBox(
-      width: MediaQuery.sizeOf(context).width / 1.2,
-      child: TextFormField(
-        controller: controller,
-        validator: (value) {
-          if (notValid && controller.text.isEmpty) {
-            return '비어 있는 일지를 만들 수 없습니다.';
-          } else {
-            return null;
-          }
-        },
-        style: textStyle,
-        decoration: const InputDecoration(
-          hintText: '제목',
-          fillColor: Colors.white,
-          disabledBorder: UnderlineInputBorder(
-            borderSide: BorderSide(
-              width: 0.5,
-            ),
-          ),
-          enabledBorder: UnderlineInputBorder(
-            borderSide: BorderSide(
-              width: 0.5,
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class _ContentForm extends ConsumerWidget {
-  const _ContentForm(
-      {super.key, required this.controller, required this.onImagePick});
-
-  final TextEditingController controller;
-
-  final void Function() onImagePick;
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final userRef = ref.watch(userControllerProvider);
-    AsyncValue<Widget> tag = userRef.whenData((appUser) {
-      if (appUser!.plants.isNotEmpty) {
-        return FittedBox(
-          fit: BoxFit.fitWidth,
-          child: Text.rich(
-            overflow: TextOverflow.ellipsis,
-            maxLines: null,
-            TextSpan(
-              text: '${appUser.plants.first.name}, ',
-              style: TextStyle(
-                color: Theme.of(context).primaryColor,
-                fontWeight: FontWeight.bold,
-              ),
-              children: [
-                TextSpan(
-                  text: appUser.plants.first.place,
-                  style: const TextStyle(
-                    color: Colors.black,
-                    fontWeight: FontWeight.normal,
-                  ),
-                ),
-              ],
-            ),
-          ),
-        );
-      } else {
-        return const SizedBox.shrink();
-      }
-    });
-    return SizedBox(
-      width: MediaQuery.sizeOf(context).width / 1.2,
-      height: MediaQuery.sizeOf(context).height / 3.1,
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          tag.isLoading
-              ? const SizedBox.shrink()
-              : Padding(
-                  padding: const EdgeInsets.only(left: 12),
-                  child: tag.value!,
-                ),
-          Expanded(
-            child: Stack(
-              children: [
-                TextFormField(
-                  controller: controller,
-                  decoration: const InputDecoration(
-                    hintText: '글쓰기 시작...',
-                    fillColor: Colors.white,
-                    disabledBorder: UnderlineInputBorder(
-                      borderSide: BorderSide.none,
-                    ),
-                    enabledBorder: UnderlineInputBorder(
-                      borderSide: BorderSide.none,
-                    ),
-                    focusedBorder: UnderlineInputBorder(
-                      borderSide: BorderSide.none,
-                    ),
-                  ),
-                  minLines: 5,
-                  maxLines: null,
-                ),
-                Align(
-                  alignment: Alignment.bottomRight,
-                  child: IconButton(
-                    onPressed: () {
-                      onImagePick();
-                    },
-                    icon: const Icon(Icons.camera_alt_rounded, size: 40),
-                  ),
-                ),
-              ],
             ),
           ),
         ],
